requirements.txt
selenium==4.1.0
pyyaml==6.0
ddt==1.5.0